// src/_workspace/types/cost-condition/IndirectCostCondition.ts
export interface IndirectCostConditionI {
  INDIRECT_COST_CONDITION_ID: number
  PRODUCT_MAIN_ID: number
  PRODUCT_MAIN_NAME: string
  LABOR: number
  DEPRECIATION: number
  OTHER_EXPENSE: number
  TOTAL_INDIRECT_COST: number
  FISCAL_YEAR: string
  VERSION: number
  UPDATE_DATE: string
  UPDATE_BY: string
}